// iot-service/iot-scada/src/sdk/models/scadaDiagramListItemResponseDto.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * IoT Monitoring System API
 * Complete IoT monitoring system REST API with nested data queries, aggregations, and real-time telemetry. Supports CRUD operations for all entities, complex queries with relations, and custom reports for dashboards.
 * OpenAPI spec version: 1.0.0
 */
import type { ScadaDiagramListItemResponseDtoProjectId } from './scadaDiagramListItemResponseDtoProjectId';
import type { ScadaDiagramListItemResponseDtoDescription } from './scadaDiagramListItemResponseDtoDescription';

export interface ScadaDiagramListItemResponseDto {
  id: string;
  ownerId: string;
  projectId?: ScadaDiagramListItemResponseDtoProjectId;
  name: string;
  description?: ScadaDiagramListItemResponseDtoDescription;
  status: string;
  updatedAt: string;
  nodeCount?: number;
  edgeCount?: number;
}